requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-httpauth==4.8.0
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
pillow==11.0.0
gunicorn==23.0.0